refactor(snapshots): to their own crate
A revival of #10472, which is generally cleaner and facilitates cross-compiling deno, thus unblocking arm64 CI builds
