+++ cli/Cargo.toml
@@ -25,17 +25,7 @@ harness = false
 path = "./bench/lsp_bench_standalone.rs"
 
 [build-dependencies]
-deno_broadcast_channel = { version = "0.49.0", path = "../ext/broadcast_channel" }
-deno_console = { version = "0.55.0", path = "../ext/console" }
-deno_core = { version = "0.137.0", path = "../core" }
-deno_crypto = { version = "0.69.0", path = "../ext/crypto" }
-deno_fetch = { version = "0.78.0", path = "../ext/fetch" }
-deno_net = { version = "0.47.0", path = "../ext/net" }
-deno_url = { version = "0.55.0", path = "../ext/url" }
-deno_web = { version = "0.86.0", path = "../ext/web" }
-deno_webgpu = { version = "0.56.0", path = "../ext/webgpu" }
-deno_websocket = { version = "0.60.0", path = "../ext/websocket" }
-deno_webstorage = { version = "0.50.0", path = "../ext/webstorage" }
+deno_runtime = { version = "0.63.0", path = "../runtime" }
 regex = "=1.5.5"
 serde = { version = "=1.0.136", features = ["derive"] }
 zstd = '=0.11.1'
@@ -52,6 +42,7 @@ deno_emit = "0.2.0"
 deno_graph = "0.27.0"
 deno_lint = { version = "0.30.0", features = ["docs"] }
 deno_runtime = { version = "0.63.0", path = "../runtime" }
+deno_snapshots = { version = "0.0.0", path = "../snapshots" }
 deno_task_shell = "0.3.1"
 
 atty = "=0.2.14"
