+++ cli/standalone.rs
@@ -307,6 +307,7 @@ pub async fn run(
     shared_array_buffer_store: None,
     compiled_wasm_module_store: None,
     stdio: Default::default(),
+    startup_snapshot: Some(deno_snapshots::cli_snapshot()),
   };
   let mut worker = MainWorker::bootstrap_from_options(
     main_module.clone(),
