+++ runtime/Cargo.toml
@@ -21,25 +21,6 @@ path = "lib.rs"
 name = "hello_runtime"
 path = "examples/hello_runtime.rs"
 
-[build-dependencies]
-deno_broadcast_channel = { version = "0.49.0", path = "../ext/broadcast_channel" }
-deno_console = { version = "0.55.0", path = "../ext/console" }
-deno_core = { version = "0.137.0", path = "../core" }
-deno_crypto = { version = "0.69.0", path = "../ext/crypto" }
-deno_fetch = { version = "0.78.0", path = "../ext/fetch" }
-deno_ffi = { version = "0.42.0", path = "../ext/ffi" }
-deno_http = { version = "0.49.0", path = "../ext/http" }
-deno_net = { version = "0.47.0", path = "../ext/net" }
-deno_tls = { version = "0.42.0", path = "../ext/tls" }
-deno_url = { version = "0.55.0", path = "../ext/url" }
-deno_web = { version = "0.86.0", path = "../ext/web" }
-deno_webgpu = { version = "0.56.0", path = "../ext/webgpu" }
-deno_webidl = { version = "0.55.0", path = "../ext/webidl" }
-deno_websocket = { version = "0.60.0", path = "../ext/websocket" }
-deno_webstorage = { version = "0.50.0", path = "../ext/webstorage" }
-
-lzzzz = '=0.8.0'
-
 [target.'cfg(windows)'.build-dependencies]
 winres = "0.1.11"
 winapi = "0.3.9"
