+++ runtime/web_worker.rs
@@ -337,6 +337,7 @@ pub struct WebWorkerOptions {
   pub compiled_wasm_module_store: Option<CompiledWasmModuleStore>,
   pub maybe_exit_code: Option<Arc<AtomicI32>>,
   pub stdio: Stdio,
+  pub startup_snapshot: Option<deno_core::Snapshot>,
 }
 
 impl WebWorker {
@@ -433,6 +434,8 @@ impl WebWorker {
       ops::tty::init(),
       deno_http::init(),
       ops::http::init(),
+      // Runtime JS
+      js::init(),
       // Permissions ext (worker specific state)
       perm_ext,
     ];
@@ -442,7 +445,7 @@ impl WebWorker {
 
     let mut js_runtime = JsRuntime::new(RuntimeOptions {
       module_loader: Some(options.module_loader.clone()),
-      startup_snapshot: Some(js::deno_isolate_init()),
+      startup_snapshot: options.startup_snapshot.clone(),
       source_map_getter: options.source_map_getter,
       get_error_class_fn: options.get_error_class_fn,
       shared_array_buffer_store: options.shared_array_buffer_store.clone(),
