+++ snapshots/Cargo.toml
@@ -1,0 +1,29 @@
+# Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
+
+[package]
+name = "deno_snapshots"
+version = "0.0.0"
+authors = ["the Deno authors"]
+edition = "2018"
+license = "MIT"
+repository = "https://github.com/denoland/deno"
+description = "Provides snapshots of TSC & Deno (runtime+web+core)"
+
+[lib]
+name = "deno_snapshots"
+path = "lib.rs"
+
+[dependencies]
+deno_core = { version = "0.137.0", path = "../core" } # For mock TSC #[op]s
+deno_runtime = { version = "0.63.0", path = "../runtime" }
+lzzzz = '=0.8.0'
+once_cell = "=1.10.0"
+zstd = '=0.11.1'
+
+[build-dependencies]
+deno_core = { version = "0.137.0", path = "../core" } # For mock TSC #[op]s
+deno_runtime = { version = "0.63.0", path = "../runtime" }
+lzzzz = '=0.8.0'
+regex = "=1.5.5"
+serde = { version = "1.0.125", features = ["derive"] }
+zstd = '=0.11.1'
